+++ CHANGELOG.rst
@@ -6,8 +6,10 @@ are documented in this file.
 
 v3.3.2
 -----------------
+* Print correct coordinates and values on plots.
+* Set time axis to numpy's datetime64 format.
 * Moved from Matplotlib style seaborn-darkgrid to seaborn-dark.
-* Grid drawn by hand in contour plot. Set automatic grid to
+* Grid is drawn by hand in contour plot. Set automatic grid to
   False: `self.axes.grid(False)`.
 
 v3.3.1 (Feb 2021)